requirements.txt
Flask==1.0.2
requests==2.21.0
gunicorn==19.9.0
python-dotenv==0.10.1
pytest==4.3.0